finalised anchors and positioning
Final setup of this stage of the project. All the anchors were repositioned so that the test file is successfully created without obvious clashes. There is one character u1CF24 that alters it's position in a number of ways. I was not able to come up with a logical algorythm to use contextual alternates, ligatures or anything that would make sense to make it work the same way as in the original file.